+++ templates/xray.html
@@ -39,10 +39,10 @@
                 </table>
         </div>
         <div class="section">
-            <img src="/static/uploads/input.jpg" alt="">
-            <img src="/static/uploads/riskmeter.jpg" alt="">
-            <img src="/static/uploads/comparison.jpg" alt="">
-            <img src="/static/uploads/rateacc.jpg" alt="">
+            <img src={{content.input.}} alt="">
+            <img src=/static/outputs/bar_graph.png alt="">
+            <img src="/static/outputs/heatmap.png" alt="">
+            <img src="/static/outputs/speedometers.png" alt="">
         </div>
     </div>
     <div class="button-container">
